pyproject.toml
[project]
name = "mcp-web-tools"
version = "0.7.5"
description = "A powerful MCP server to equip LLMs with web access, search, and content extraction capabilities"
readme = "README.md"
authors = [
    { name = "Paul-Louis Pröve", email = "mail@plpp.de" }
]
license = { text = "MIT" }
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing :: Markup",
]
keywords = ["mcp", "llm", "web", "search", "extraction", "pdf"]
dependencies = [
    "duckduckgo-search>=8.1.1",
    "googlesearch-python>=1.3.0",
    "mcp>=1.13.1",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pymupdf4llm>=0.0.27",
    "trafilatura>=2.0.0",
    "zendriver>=0.14.0",
    "brave-search-python-client>=0.4.27",
]

[project.urls]
Homepage = "https://github.com/pietz/mcp-web-tools"
Issues = "https://github.com/pietz/mcp-web-tools/issues"

[project.scripts]
mcp-web-tools = "mcp_web_tools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
